pyproject.toml
[project]
name = "api-service"
version = "0.1.0"
description = "MotoKönig - Social network for Kaliningrad motorcyclists"
requires-python = ">=3.12"
dependencies = [
    "advanced-alchemy>=1.4.2",
    "aioboto3>=14.3.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.15.2",
    "bcrypt>=4.3.0",
    "dishka>=1.5.3",
    "fastapi[standard]>=0.115.12",
    "faststream[rabbit]>=0.5.40",
    "passlib>=1.7.4",
    "pyjwt[crypto]>=2.10.1",
    "redis>=5.2.1",
    "ruff>=0.11.9",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio<1.0.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "faker>=33.2.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long
    "B008", # do not perform function calls in argument defaults
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
